copy task within run loop of delegate method
The delegate method `urlSession(_ session: URLSession, downloadTask: URLSessionDownloadTask, didFinishDownloadingTo location: URL)` requires that the file be copied before the end of the method is reached. The file at the given location URL is deleted immediately after. Therefore an `async` execution of the task change does not copy the file in time leading to empty files.